safari-600.7-branch/LayoutTests/fast/viewport/viewport-68-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "100x" for key "width" was truncated to its numeric prefix.
ALERT: viewport size 320x352 scale 1 with limits [1, 5] and userScalable true
